Add files via upload
This is a declaration of a deployment and service (based on the previous files). With a demo-ingress.yml, I'll be able to demonstrate how the ingress works. This can show the appropriate pod's index file when a user requests http://domainname/demo in url without they type any port number at the end.